Rename header and cpp files to match new class name (OP2SpriteLoader)
