+++ OP2Utility.vcxproj.filters
@@ -126,9 +126,6 @@
     <ClInclude Include="src\Sprite\Animation.h">
       <Filter>Sprite</Filter>
     </ClInclude>
-    <ClInclude Include="src\Sprite\OP2BmpLoader.h">
-      <Filter>Sprite</Filter>
-    </ClInclude>
     <ClInclude Include="src\Bitmap\BmpCompression.h">
       <Filter>Bitmap</Filter>
     </ClInclude>
@@ -179,6 +176,9 @@
     </ClInclude>
     <ClInclude Include="src\Stream\DynamicMemoryWriter.h">
       <Filter>Stream</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Sprite\OP2SpriteLoader.h">
+      <Filter>Sprite</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -251,9 +251,6 @@
     <ClCompile Include="src\Sprite\SectionHeader.cpp">
       <Filter>Sprite</Filter>
     </ClCompile>
-    <ClCompile Include="src\Sprite\OP2BmpLoader.cpp">
-      <Filter>Sprite</Filter>
-    </ClCompile>
     <ClCompile Include="src\Archive\WaveFile.cpp">
       <Filter>Archive</Filter>
     </ClCompile>
@@ -283,6 +280,9 @@
     </ClCompile>
     <ClCompile Include="src\Map\SavedGameUnits.cpp">
       <Filter>Map</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Sprite\OP2SpriteLoader.cpp">
+      <Filter>Sprite</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
