+++ src/Sprite/OP2SpriteLoader.cpp
@@ -1,4 +1,4 @@
-#include "OP2BmpLoader.h"
+#include "OP2SpriteLoader.h"
 #include "../Bitmap/ImageHeader.h"
 #include "../Bitmap/BitmapFile.h"
 #include <stdexcept>
