+++ src/Sprite/OP2SpriteLoader.h
